+++ lib/locatine/for_search/saying.rb
@@ -91,8 +91,23 @@ module Locatine
         push_title "#{length} #{verb(length)} selected as #{name} in #{scope}."\
         " But it was already defined #{found.length} times."
         example = found.sample
+        same = found.select { |i| (i[:name] == name) && (i[:scope] == scope) }
+        if same.empty?
+          send_same_entry_example(example)
+        else
+          send_fully_similar
+        end
+      end
+
+      def send_same_entry_example(example)
         send_to_app('locatinehint', "For example like #{example[:name]} in"\
         " #{example[:scope]}")
+      end
+
+      def send_fully_similar
+        send_to_app('locatinehint', 'And it was defined with the same name'\
+        ' and in the same scope. So saving your selection will cause'\
+        ' deleting of all previously stored data about the element')
       end
 
       def warn_dropping(tag, index)