+++ spec/e2e_spec.rb
@@ -81,9 +81,9 @@ describe 'E2E basic user story' do
     expect(@s.browser.element(@s.lctr(name: "important span", vars: {"tag":"spoon", "text":"kedavra", "var":"ba"})).text).to be == "Abrakedavra"
   end
 
-  it "Ignores exact if element still unstable" do
+  it "Fails when there is no exact" do
     @s.browser.goto @path2
-    expect(@s.find(name: "span for guess", exact: true).text).to be == "for guess"
+    expect{@s.check(name: "span for guess", exact: true)}.to raise_error(RuntimeError, 'Locatine cannot find element span for guess in Default')
   end
 
   it "Finding exacts" do
@@ -94,20 +94,23 @@ describe 'E2E basic user story' do
 
   it "Fails on exact when elements are lost" do
     @s.browser.goto @path3
-    expect(@s.collect(name: "lis fox", exact: true, locator: {id: "not welcome"})).to be == nil
-    expect(@s.collect(name: "lis fox", exact: true, tolerance: 0)).to be == nil
-    @s.stability_limit = 7
-    expect(@s.collect(name: "lis fox", exact: true)).to be == nil
-    @s.stability_limit = 4
-    expect(@s.find(name: "element", exact: true)).to be == nil
-    @s.stability_limit = 2
-    expect(@s.find(name: "span for guess", exact: true)).to be == nil
+    expect(@s.collect(name: "lis fox", exact: true, no_fail: true, locator: {id: "not welcome"})).to be == []
+    expect(@s.check_collection(name: "lis fox", exact: true, no_fail: true)).to be == []
+    expect(@s.exact_collection(name: "lis fox", no_fail: true)).to be == []
+    expect(@s.exact(name: "element", no_fail: true)).to be == nil
+    expect(@s.find(name: "span for guess", no_fail: true, exact: true)).to be == nil
   end
 
   it "Fails when elements are lost and there is nothing similar" do
     @s.browser.goto @path4
     expect{@s.collect(name: "lis fox")}.to raise_error(RuntimeError, 'Locatine cannot find element lis fox in Default')
     expect{@s.find(name: "element")}.to raise_error(RuntimeError, 'Locatine cannot find element similar to element in Default')
+  end
+
+  it "Will not fail with no_fail == true" do
+    @s.browser.goto @path4
+    expect(@s.collect(name: "lis fox", no_fail: true)).to be == []
+    expect(@s.find(name: "element", no_fail: true)).to be == nil
   end
 
   it "Fails when visual is turned off but search is very css related" do